edit logic get item_blance all method

+++ database/migrations/2021_10_26_084516_create_item_transactions_table.php
@@ -45,6 +45,9 @@ class CreateItemTransactionsTable extends Migration
      */
     public function down()
     {
+        if(app()->isProduction())
+            return;
+
         Schema::dropIfExists('item_transactions');
     }
 }
